requirements.txt
# Core web framework
streamlit>=1.28.0

# Authentication and cookies
streamlit-cookies-manager>=0.2.0
pyrebase4>=4.7.1

# AI/ML libraries
torch>=2.0.0
transformers>=4.30.0
spacy>=3.6.0
scikit-learn>=1.3.0

# NLP and text processing
nltk>=3.8.1
python-dateutil>=2.8.2

# PDF processing
PyPDF2>=3.0.1

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Additional utilities
requests>=2.31.0
urllib3>=1.26.0
